requirements.txt
torch
pandas
numpy
scikit-learn
pytest
shap
python-docx
matplotlib
pyyaml
fastapi
pydantic
uvicorn
seaborn